refactoring react files to including react router and react bootstrap. Added Home component
